requirements.txt
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
numpy==1.25.2
pandora==0.2.0
scikitlearn==1.3.0
pillow==10.0.0
tensorflow==2.13.0
flask-cors==4.0.0